requirements.txt
python-telegram-bot>=21.0,<22
anthropic>=0.40.0,<1
supabase>=2.0.0,<3
httpx>=0.27.0,<1
